requirements.txt
asgiref==3.2.10
autopep8==1.5.3
Django==3.0.6
django-cors-headers==3.4.0
django-filter==2.3.0
djangorestframework==3.11.0
importlib-metadata==1.7.0
Markdown==3.2.2
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
zipp==3.1.0
